feat(x86_64): 重构信号处理器与蹦床机制，支持分支指令重写
重构 x86_64 架构的信号处理机制，从原有的单步模式改为自包含蹦床方案：
- 移除 EFLAGS.TF 单步依赖，改为一次 SIGTRAP + 蹦床内嵌回跳逻辑
- 支持 CALL/JMP 相对分支指令的绝对地址重写
- 新增 Jcc 条件分支的反条件跳转+绝对跳转重写方案
- 添加 x86_64 JCC 和 aarch64 BL 的汇编测试用例
- 更新方案文档，详细说明新的蹦床机制和分支处理逻辑

重构提高了执行效率（减少一次信号陷阱）并扩展了对控制流指令的支持范围。

+++ src/arch/aarch64.rs
@@ -114,7 +114,7 @@ fn save_old_instruction(
             // BR X17, for B, B.cond etc.
             new_bytes.extend_from_slice(&0xd61f0220_u32.to_le_bytes());
         }
-        new_bytes.extend_from_slice(&target_addr.to_le_bytes());
+        new_bytes.extend_from_slice(&(target_addr as u64).to_le_bytes());
 
         let old_len = ins.bytes().len();
         let new_len = new_bytes.len();
